refactor: slightly simplify
