+++ package.json
@@ -118,8 +118,9 @@
     ]
   },
   "scripts": {
+    "generate:contributes": "jiti ./scripts/generate-contributes.ts",
     "pretest": "nr build && nr lint",
-    "build": "nr ui:build && rollup -c",
+    "build": "nr generate:contributes && nr ui:build && rollup -c",
     "dev": "nr build --watch --sourcemap",
     "ui:dev": "nuxt dev ui",
     "ui:build": "nuxt build ui",
