--- scripts/updatePackageJson.ts
@@ -1,11 +1,0 @@
-import { readFileSync, writeFileSync } from 'node:fs'
-import { contributesForPackageJson } from '../src/genPackageContributions'
-
-const str = readFileSync('./package.json').toString()
-const json = JSON.parse(str)
-
-json.contributes = contributesForPackageJson
-
-const outStr = JSON.stringify(json, null, 2)
-
-writeFileSync('./package.json', outStr)
